do not overwrite or rename files when filenames clash

+++ mnn/plotting/utils.py
@@ -1,3 +1,4 @@
+import logging
 import os
 
 
@@ -40,7 +41,12 @@ def save_fig(fig, name: str):
     dir_name = "plots"
     os.makedirs(dir_name, exist_ok=True)
     path = os.path.join(dir_name, f"{name}.pdf")
+    if os.path.exists(path):
+        logging.warning(f'File "{path}" already exists. Skipping...')
+        return
+
     fig.savefig(path, bbox_inches="tight", transparent=True)
+    logging.info(f'Saved file "{path}".')
 
 
 def add_boxplot_legend(axis, boxplots, labels, linewdith=1.0, loc="upper right"):
